Refactoring
Refactoring SpiderBot.cs to remove functions that can be reused into EnemyController.cs, which will be a parent to all enemy scripts.
