+++ Assets/Prefabs/Enemy.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 1704585594299060261}
   - component: {fileID: 1704585594299060282}
   - component: {fileID: 1704585594299060283}
+  - component: {fileID: 1503021785595427093}
   - component: {fileID: 1704585594299060280}
   m_Layer: 0
   m_Name: Enemy
@@ -83,6 +84,22 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
+--- !u!114 &1503021785595427093
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1704585594299060281}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d63da978b6f61494680b63f10cd7359c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 3
+  attackDamage: 1
+  speed: 2.5
+  facingRight: 1
 --- !u!114 &1704585594299060280
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -98,9 +115,8 @@ MonoBehaviour:
   maxHp: 3
   attackDamage: 1
   speed: 2.5
-  gcRaycastDist: 1
   facingRight: 1
-  walkOnCeiling: 0
+  startingPosition: 0
   whatIsGround:
     serializedVersion: 2
     m_Bits: 64
