+++ Assets/Scripts/scr_Enemies/EnemyController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 215dac9d201ec884ebd0c6b61d2786ad
+guid: d63da978b6f61494680b63f10cd7359c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
